+++ src/lib/codec-importer.ts
@@ -4,34 +4,34 @@ import { type BlockCodec } from 'multiformats/codecs/interface'
 import getCodecNameFromCode from './get-codec-name-from-code'
 
 type CodecDataTypes = PBNode | Uint8Array
-interface CodecImporterResponse<T> extends Pick<BlockCodec<number, T | unknown>, 'decode'> {
+interface CodecImporterResponse<T> extends Pick<BlockCodec<number, T | unknown>, 'decode' | 'encode' | 'code'> {
 }
 
-export default async function codecImporter<T extends CodecDataTypes = CodecDataTypes> (codecCode: number): Promise<CodecImporterResponse<T>> {
-  const codecName = getCodecNameFromCode(codecCode)
+export default async function codecImporter<T extends CodecDataTypes = CodecDataTypes> (codeOrName: number | string): Promise<CodecImporterResponse<T>> {
+  let codecName: string
+  if (typeof codeOrName === 'string') {
+    codecName = codeOrName
+  } else {
+    codecName = getCodecNameFromCode(codeOrName)
+  }
+
+  // #WhenAddingNewCodec
   switch (codecName) {
     case 'dag-cbor':
       return import('@ipld/dag-cbor')
     case 'dag-pb':
-      // @ts-expect-error - return types need normalizing
       return import('@ipld/dag-pb')
     case 'git-raw':
-      return {
-        decode: (await import('ipld-git')).default.util.deserialize
-      }
+      throw new Error('git-raw is unsupported until https://github.com/ipld/js-ipld-git is updated.')
     case 'raw':
-      // @ts-expect-error - return types need normalizing
       return import('multiformats/codecs/raw')
     case 'json':
       return import('multiformats/codecs/json')
     case 'dag-json':
       return import('@ipld/dag-json')
     case 'dag-jose':
-      return {
-        decode: (await import('dag-jose')).decode as unknown as
-          ((bytes: Uint8Array) => Promise<unknown>)
-      }
+      return import('dag-jose')
     default:
-      throw new Error(`unsupported codec: ${codecCode}=${getCodecNameFromCode(codecCode)}`)
+      throw new Error(`unsupported codec: ${codeOrName}=${codecName}`)
   }
 }