+++ src/lib/init-helia.ts
@@ -1,24 +1,34 @@
 import { noise } from '@chainsafe/libp2p-noise'
 import { yamux } from '@chainsafe/libp2p-yamux'
+import { createDelegatedRoutingV1HttpApiClient } from '@helia/delegated-routing-v1-http-api-client'
 import { type Helia } from '@helia/interface'
-import { delegatedContentRouting } from '@libp2p/delegated-content-routing'
-import { delegatedPeerRouting } from '@libp2p/delegated-peer-routing'
-import { mplex } from '@libp2p/mplex'
+import { autoNAT } from '@libp2p/autonat'
+import { circuitRelayTransport } from '@libp2p/circuit-relay-v2'
+import { identify } from '@libp2p/identify'
 import { webRTC, webRTCDirect } from '@libp2p/webrtc'
 import { webSockets } from '@libp2p/websockets'
 import { webTransport } from '@libp2p/webtransport'
 import { MemoryBlockstore } from 'blockstore-core'
 import { MemoryDatastore } from 'datastore-core'
 import { createHelia } from 'helia'
-import { type create as createKuboClient } from 'kubo-rpc-client'
+import { trustlessGateway } from 'helia/block-brokers'
 import { createLibp2p } from 'libp2p'
-import { autoNATService } from 'libp2p/autonat'
-import { circuitRelayTransport } from 'libp2p/circuit-relay'
-import { identifyService } from 'libp2p/identify'
 
-import { addDagNodeToHelia } from '../lib/helpers'
+import { getHashersForCodes } from './hash-importer.js'
+import { addDagNodeToHelia } from '../lib/helpers.js'
+import type { KuboGatewayOptions } from '../types.d.js'
 
-export default async function initHelia (kuboClient: ReturnType<typeof createKuboClient>): Promise<Helia> {
+function areRemoteGatewaysEnabled (): boolean {
+  const localStorageKey = 'explore.ipld.gatewayEnabled'
+  console.info(
+    `🎛️ Customise whether ipld-explorer-components fetches content from gateways by setting an '${localStorageKey}' value to true/false in localStorage. e.g. localStorage.setItem('explore.ipld.gatewayEnabled', false) -- NOTE: defaults to true`
+  )
+  const gatewayEnabledSetting = localStorage.getItem(localStorageKey)
+
+  return gatewayEnabledSetting != null ? JSON.parse(gatewayEnabledSetting) : true
+}
+
+export default async function initHelia (kuboGatewayOptions: KuboGatewayOptions): Promise<Helia> {
   const blockstore = new MemoryBlockstore()
   const datastore = new MemoryDatastore()
 
@@ -26,13 +36,10 @@ export default async function initHelia (kuboClient: ReturnType<typeof createKub
    * based on https://github.com/ipfs/helia/blob/ed4985677b62021f76541354ad06b70bd53e929a/packages/helia/src/utils/libp2p.browser.ts#L20
    */
   const libp2p = await createLibp2p({
-    start: true, // TODO: libp2p bug with stop/start - https://github.com/libp2p/js-libp2p/issues/1787
-    peerRouters: [
-      delegatedPeerRouting(kuboClient)
-    ],
-    contentRouters: [
-      delegatedContentRouting(kuboClient)
-    ],
+    connectionManager: {
+      // do not auto-dial peers. We will manually dial peers when we need them.
+      minConnections: 0
+    },
     datastore,
     transports: [
       webRTC(),
@@ -47,23 +54,46 @@ export default async function initHelia (kuboClient: ReturnType<typeof createKub
       noise()
     ],
     streamMuxers: [
-      yamux(),
-      mplex()
+      yamux()
     ],
     services: {
-      identify: identifyService(),
-      autoNAT: autoNATService()
+      identify: identify(),
+      autoNAT: autoNAT(),
+      delegatedRouting: () => createDelegatedRoutingV1HttpApiClient('https://delegated-ipfs.dev')
     }
   })
 
+  // Always add the Kubo gatewawy
+  const trustlessGateways = [
+    trustlessGateway({ gateways: [`${kuboGatewayOptions.protocol ?? 'http'}://${kuboGatewayOptions.host}:${kuboGatewayOptions.port}`] })
+  ]
+
+  if (areRemoteGatewaysEnabled()) {
+    trustlessGateways.push(trustlessGateway())
+  }
+
   const helia = await createHelia({
+    blockBrokers: [
+      // no bitswap
+      ...trustlessGateways
+    ],
+    // #WhenAddingNewHasher
+    hashers: await getHashersForCodes(17, 18, 19, 20, 27, 30),
     datastore,
     blockstore,
+    // @ts-expect-error - libp2p types are borked right now
     libp2p
   })
 
   // add helia-only examples
-  await addDagNodeToHelia(helia, await import('@ipld/dag-json'), { hello: 'world' }) // baguqeerasords4njcts6vs7qvdjfcvgnume4hqohf65zsfguprqphs3icwea
+  // consumers may not have the peer-deps installed for these examples, and we don't want to break them if they're not supported.
+  await Promise.allSettled([
+    addDagNodeToHelia(helia, 'dag-json', { hello: 'world' }), // baguqeerasords4njcts6vs7qvdjfcvgnume4hqohf65zsfguprqphs3icwea
+    addDagNodeToHelia(helia, 'dag-cbor', { hello: 'world' }, 27), // bafyrwigbexamue2ba3hmtai7hwlcmd6ekiqsduyf5avv7oz6ln3radvjde
+    addDagNodeToHelia(helia, 'json', { hello: 'world' }, 20), // bagaaifcavabu6fzheerrmtxbbwv7jjhc3kaldmm7lbnvfopyrthcvod4m6ygpj3unrcggkzhvcwv5wnhc5ufkgzlsji7agnmofovc2g4a3ui7ja
+    addDagNodeToHelia(helia, 'json', { hello: 'world' }, 30), // bagaaihraf4oq2kddg6o5ewlu6aol6xab75xkwbgzx2dlot7cdun7iirve23a
+    addDagNodeToHelia(helia, 'raw', (new TextEncoder()).encode('hello'), 30) // bafkr4ihkr4ld3m4gqkjf4reryxsy2s5tkbxprqkow6fin2iiyvreuzzab4
+  ])
 
   return helia
 }